requirements.txt
beautifulsoup4
blinker==1.6.2
celery==5.2.7
django-redis
djangorestframework
dnspython
drf-spectacular
httpx
joblib
networkx
tldextract
sastrawi
scikit-learn
selenium-wire
pandas
python-dotenv
python-whois
reportlab
serpapi
search-results
tldextract